Use kT = 0.1 for low temperature part
The simulation at kT = 0.5 showed a decrease in the potential energy,
but a very rapid ad small decrease. Simulating at kT = 0.1 makes the
system approach the lower energy more gradfually, which makes it more
clear to the reader that something is actually changing in the system
when the temperature changes.